requirements.txt
requests~=2.24.0
atlassian-python-api~=2.0.1
